+++ src/bmf_collapse.h
