+++ src/bmf_main.cpp
@@ -32,9 +32,9 @@ int main(int argc, char *argv[])
         exit(EXIT_SUCCESS);
     }
     if(strcmp(argv[1], "sort") == 0) return sort_main(argc - 1, argv + 1);
-    if(strcmp(argv[1], "collapse") == 0) return bmf::dmp_main(argc - 1, argv + 1);
+    if(strcmp(argv[1], "collapse") == 0) return bmf::collapse_main(argc - 1, argv + 1);
     if(strcmp(argv[1], "rsq") == 0) return bmf::rsq_main(argc - 1, argv + 1);
-    if(strcmp(argv[1], "hashdmp") == 0) return bmf::hashdmp_main(argc - 1, argv + 1);
+    if(strcmp(argv[1], "hashdmp") == 0) return bmf::hashcollapse_main(argc - 1, argv + 1);
     if(strcmp(argv[1], "inmem") == 0) return bmf::hashdmp_inmem_main(argc - 1, argv + 1);
     if(strcmp(argv[1], "famstats") == 0) return bmf::famstats_main(argc - 1, argv + 1);
     if(strcmp(argv[1], "vet") == 0) return bmf::vet_main(argc - 1, argv + 1);
@@ -45,6 +45,14 @@ int main(int argc, char *argv[])
     if(strcmp(argv[1], "depth") == 0) return bmf::depth_main(argc - 1, argv + 1);
     if(strcmp(argv[1], "stack") == 0) return bmf::stack_main(argc - 1, argv + 1);
     if(strcmp(argv[1], "filter") == 0) return bmf::filter_main(argc - 1, argv + 1);
+    if(strcmp(argv[1], "dmp") == 0) {
+        LOG_WARNING("bmftools dmp has been renamed 'bmftools collapse inline'\n");
+        return bmf::idmp_main(argc - 1, argv + 1);
+    }
+    if(strcmp(argv[1], "sdmp") == 0) {
+        LOG_WARNING("bmftools sdmp has been renamed 'bmftools collapse secondary'\n");
+        return bmf::sdmp_main(argc - 1, argv + 1);
+    }
     fprintf(stderr, "Unrecognized command %s. Abort!\n", argv[1]);
     return EXIT_FAILURE;
 }
