+++ src/bmf_main.h
@@ -12,12 +12,13 @@ extern "C" {
 namespace bmf {
 extern int cap_main(int argc, char *argv[]);
 extern int depth_main(int argc, char *argv[]);
-extern int dmp_main(int argc, char *argv[]);
+extern int collapse_main(int argc, char *argv[]);
 extern int err_main(int argc, char *argv[]);
 extern int famstats_main(int argc, char *argv[]);
 extern int filter_main(int argc, char *argv[]);
-extern int hashdmp_main(int argc, char *argv[]);
+extern int hashcollapse_main(int argc, char *argv[]);
 extern int hashdmp_inmem_main(int argc, char *argv[]);
+extern int idmp_main(int argc, char *argv[]);
 extern int mark_main(int argc, char *argv[]);
 extern int rsq_main(int argc, char *argv[]);
 extern int sdmp_main(int argc, char *argv[]);
